change wheelMax to float value for math error

+++ belugapad/RatioTool.h
@@ -45,7 +45,7 @@
     
     NSMutableArray *numberWheels;
     
-    int wheelMax;
+    float wheelMax;
 }
